added conversion to uint32 to allow map display

+++ src/utils/tiles/LatLonToTileNumber.m
@@ -8,11 +8,11 @@ n           = 2^zoom;
 tile        = Tile;
 
 % X
-tile.Column = n * ((longitude + 180) / 360);
+tile.Column = uint32(n * ((longitude + 180) / 360));
 
 % Y
 lat_rad     = deg2rad(latitude);
-tile.Line   = n * (1 - (log(tan(lat_rad) + sec(lat_rad)) / pi)) /2;
+tile.Line   = uint32(n * (1 - (log(tan(lat_rad) + sec(lat_rad)) / pi)) /2);
 
 end
